requirements.txt
streamlit
feedparser
deep-translator
requests
python-dotenv
APScheduler
